Refactor item IDs and add mining cron job
Refactored item identification from Name to ItemId across inventory, shop, and related logic for consistency and future extensibility. Introduced a cron scheduler using gocron to automate mining events, including new cron package and mining task. Updated event logic to use int64 for scores, improved error handling, and made minor bug fixes and code cleanups throughout the bot and repository layers.

+++ internal/pkg/common/score.go
@@ -57,7 +57,6 @@ func formatSmallNumber(n int64) string {
 		sign = "-"
 		n = -n
 	}
-
 	mod := math.Abs(float64(n))
 	switch {
 	case mod >= 1e15:
@@ -84,13 +83,13 @@ func int64Pow(base, exp int) int64 {
 	return result
 }
 
-func (u *Score) Increase(increment uint64) {
-	u.Mantissa += int64(increment)
+func (u *Score) Increase(increment int64) {
+	u.Mantissa += increment
 	u.normalize()
 }
 
-func (u *Score) Decrease(decrement uint64) {
-	u.Mantissa -= int64(decrement)
+func (u *Score) Decrease(decrement int64) {
+	u.Mantissa -= decrement
 	u.normalize()
 }
